feat - change to use class, rename to mugunghwa
